Modernize double precision declarations in field_can.f90
Replace remaining 'double precision' declarations with 'real(dp)'.
Module already used iso_fortran_env with dp parameter defined.

Fixes #101

+++ src/field_can.f90
@@ -159,7 +159,7 @@ end subroutine init_field_can
 
 subroutine FieldCan_init(f, mu, ro0, vpar)
   type(FieldCan), intent(inout) :: f
-  double precision, intent(in), optional  :: mu, ro0, vpar
+  real(dp), intent(in), optional  :: mu, ro0, vpar
 
   if (present(mu)) then
     f%mu = mu
@@ -190,7 +190,7 @@ subroutine get_val(f, pphi)
   !
   !
   type(FieldCan), intent(inout) :: f
-  double precision, intent(in) :: pphi
+  real(dp), intent(in) :: pphi
 
   f%vpar = (pphi - f%Aph/f%ro0)/f%hph
   f%H = f%vpar**2/2d0 + f%mu*f%Bmod
@@ -207,7 +207,7 @@ subroutine get_derivatives(f, pphi)
   !
   !
   type(FieldCan), intent(inout) :: f
-  double precision, intent(in) :: pphi
+  real(dp), intent(in) :: pphi
 
   call get_val(f, pphi)
 
@@ -233,7 +233,7 @@ subroutine get_derivatives2(f, pphi)
   ! d2dpphdr, d2dpphdth, d2dpphdph, d2dpph2
   !
   type(FieldCan), intent(inout) :: f
-  double precision, intent(in) :: pphi
+  real(dp), intent(in) :: pphi
 
   call get_derivatives(f, pphi)
